feat: add initial implementation of dashboard with user authentication and configuration management
- Introduced a new dashboard module with routes for managing clients, providers, and routes.
- Implemented user authentication with an initial admin user setup.
- Added sample data creation functions for clients, providers, and routes.
- Created a new SQLite migration for the admin_users table.
- Developed a frontend dashboard application using JavaScript and Tailwind CSS.
- Added HTML templates for the dashboard and login pages.
- Implemented integration tests for configuration versioning and health check endpoints.

+++ src/auth.rs
@@ -49,7 +49,7 @@ impl AuthRegistry {
                 allowed_routes: config.allowed_routes.clone(),
                 limiter: Mutex::new(VecDeque::new()),
                 requests_per_minute: config.requests_per_minute as usize,
-                concurrency: Arc::new(Semaphore::new(config.concurrent_requests)),
+                concurrency: Arc::new(Semaphore::new(config.concurrent_requests as usize)),
             }));
         }
         Ok(Self { clients })
